+++ sdk/storage/azure-storage-common/inc/azure/storage/common/storage_exception.hpp
@@ -8,13 +8,14 @@
 #include <stdexcept>
 #include <string>
 
+#include <azure/core/exception.hpp>
 #include <azure/core/http/http.hpp>
 
 namespace Azure { namespace Storage {
 
-  struct StorageException : public std::runtime_error
+  struct StorageException : public Azure::Core::RequestFailedException
   {
-    explicit StorageException(const std::string& message) : std::runtime_error(message) {}
+    explicit StorageException(const std::string& message) : RequestFailedException(message) {}
 
     Azure::Core::Http::HttpStatusCode StatusCode = Azure::Core::Http::HttpStatusCode::None;
     std::string ReasonPhrase;