+++ sdk/storage/azure-storage-files-datalake/src/datalake_path_client.cpp
@@ -386,7 +386,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     ret.CopyStatus = std::move(result->CopyStatus);
     ret.CopyProgress = std::move(result->CopyProgress);
     ret.CopyCompletedOn = std::move(result->CopyCompletedOn);
-    ret.ExpiresOn = std::move(result->ExpiriesOn);
+    ret.ExpiresOn = std::move(result->ExpiresOn);
     ret.LastAccessedOn = std::move(result->LastAccessedOn);
     ret.FileSize = result->ContentLength;
     ret.RequestId = std::move(result->RequestId);
@@ -394,8 +394,9 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
         std::move(ret), result.ExtractRawResponse());
   }
 
-  Azure::Core::Response<Models::GetDataLakePathAccessControlResult>
-  DataLakePathClient::GetAccessControls(const GetDataLakePathAccessControlOptions& options) const
+  Azure::Core::Response<Models::GetDataLakePathAccessControlListResult>
+  DataLakePathClient::GetAccessControlList(
+      const GetDataLakePathAccessControlListOptions& options) const
   {
     Details::DataLakeRestClient::Path::GetPropertiesOptions protocolLayerOptions;
     protocolLayerOptions.Action = Models::PathGetPropertiesAction::GetAccessControl;
@@ -411,12 +412,13 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     {
       acl = Models::Acl::DeserializeAcls(result->Acl.GetValue());
     }
-    Models::GetDataLakePathAccessControlResult ret;
+    Models::GetDataLakePathAccessControlListResult ret;
     ret.ETag = std::move(result->ETag);
     ret.LastModified = std::move(result->LastModified);
     if (!acl.HasValue())
     {
-      throw std::runtime_error("Got null value returned when getting access control.");
+      throw Azure::Core::RequestFailedException(
+          "Got null value returned when getting access control.");
     }
     ret.Acls = std::move(acl.GetValue());
     if (result->Owner.HasValue())
@@ -432,7 +434,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
       ret.Permissions = result->Permissions.GetValue();
     }
     ret.RequestId = std::move(result->RequestId);
-    return Azure::Core::Response<Models::GetDataLakePathAccessControlResult>(
+    return Azure::Core::Response<Models::GetDataLakePathAccessControlListResult>(
         std::move(ret), result.ExtractRawResponse());
   }