+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -31,10 +31,12 @@
 - Removed unused type `AccountResourceType` and `PathLeaseAction`.
 - Changed all previous `LeaseDuration` members to a new type named `LeaseDurationType`.
 - `startsOn` parameter for `GetUserDelegationKey` was changed to optional.
+- Removed `PreviousContinuationToken` from `ListFileSystemsSinglePageResult`.
 
 ### Other Changes and Improvements
 
 - Changed `DataLakeFileClient::Flush`'s `endingOffset` parameter's name to `position`.
+- Renamed `DataLakePathClient::GetAccessControls` to `DataLakePathClient::GetAccessControlList`.
 - Removed unused parameters, options, results and functions in protocol layer.
 
 ## 12.0.0-beta.6 (2020-01-14)